compressor/requirements.txt
boto3>=1.34.0
pyarrow>=15.0.0
python-dotenv>=1.0.0
